+++ docs/Whats-New.md
@@ -57,9 +57,12 @@ New:
 - Script Action 92 for waiting & repeat the same new AI attack if no target was found (by FS-21)
 - Script Action 93 that modifies the Team's Trigger Weight when ends the new attack action (by FS-21)
 - Script Action 94 for picking a random script from a list (by FS-21)
-- Script Action 112 to regroup temporarily around the Team Leader (by FS-21)
 - Script Action 95 to 98 for new AI movements towards certain objects (by FS-21)
+- Script Action 111 that un-register Team success, is just the opposite effect of Action 49 (by FS-21)
+- Script Action 112 to regroup temporarily around the Team Leader (by FS-21)
 - ObjectInfo now shows current Target and AI Trigger data (by FS-21)
+
+
 
 Vanilla fixes:
 - Fixed laser drawing code to allow for thicker lasers in house color draw mode (by Kerbiter, ChrisLv_CN)